+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updated the `Signal.flatMapLatest()` transformation to allow more flexible mixing of signal types between `self` and the signal returned from `transform`.
 - Added `Signal.toogle()` method for read-write boolean signals.
 - Added `willWrite()` and `didWrite()` transformations to read write signal.
+- Added `UIControl` `valueChanged` that will signal with the latest value when the control event `.valueChanged` is signaled.
 
 # 1.3.1